섹션 5.헥사고날 아키텍처로 회원 애플리케이션 개발
28. 회원 애플리케이션 서비스 테스트(2) - 중간 15:37
