+++ src/main/java/tobyspring/splearn/domain/DuplicateEmailException.java
@@ -1,0 +1,7 @@
+package tobyspring.splearn.domain;
+
+public class DuplicateEmailException extends RuntimeException {
+    public DuplicateEmailException(String message) {
+
+    }
+}
