+++ src/test/java/tobyspring/splearn/SplearnTestConfiguration.java
@@ -1,0 +1,29 @@
+package tobyspring.splearn;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import tobyspring.splearn.application.required.EmailSender;
+import tobyspring.splearn.domain.Email;
+import tobyspring.splearn.domain.MemberFixture;
+import tobyspring.splearn.domain.PasswordEncoder;
+
+@TestConfiguration
+public class SplearnTestConfiguration {
+
+    @Bean
+    public EmailSender emailSender() {
+        return new EmailSender() {
+            @Override
+            public void send(Email email, String subject, String body) {
+                System.out.println("Send email: " + email);
+            }
+        };
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return MemberFixture.createPasswordEncoder();
+    }
+
+
+}
