+++ src/test/java/tobyspring/splearn/application/provided/MemberRegisterTest.java
@@ -2,100 +2,37 @@ package tobyspring.splearn.application.provided;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.springframework.test.util.ReflectionTestUtils;
-import tobyspring.splearn.application.MemberService;
-import tobyspring.splearn.application.required.EmailSender;
-import tobyspring.splearn.application.required.MemberRepository;
-import tobyspring.splearn.domain.Email;
-import tobyspring.splearn.domain.Member;
-import tobyspring.splearn.domain.MemberFixture;
-import tobyspring.splearn.domain.MemberStatus;
-
-import java.util.ArrayList;
-import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import tobyspring.splearn.SplearnTestConfiguration;
+import tobyspring.splearn.domain.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class MemberRegisterTest {
+@SpringBootTest
+@Import({SplearnTestConfiguration.class})
+public class MemberRegisterTest {
+
+    @Autowired
+    private MemberRegister memberRegister;
 
     @Test
-    @DisplayName("회원 생성 Stub")
-    void registerTestStub() {
-        MemberRegister register = new MemberService(
-                new MemberRepositoryStub(), new EmailSenderStub(), MemberFixture.createPasswordEncoder()
-        );
+    @DisplayName("회원 등록")
+    void register() {
+        Member member = memberRegister.register(MemberFixture.createMemberRegisterReques());
 
-        Member member = register.register(MemberFixture.createMemberRegisterReques());
         assertThat(member.getId()).isNotNull();
         assertThat(member.getStatus()).isEqualTo(MemberStatus.PENDING);
 
     }
 
     @Test
-    @DisplayName("회원 생성 Mock")
-    void registerTestMock() {
-        EmailSenderMock emailSender = new EmailSenderMock();
-        MemberRegister register = new MemberService(
-                new MemberRepositoryStub(), emailSender, MemberFixture.createPasswordEncoder()
-        );
-
-        Member member = register.register(MemberFixture.createMemberRegisterReques());
-        assertThat(member.getId()).isNotNull();
-        assertThat(member.getStatus()).isEqualTo(MemberStatus.PENDING);
-
-        assertThat(emailSender.getTos()).hasSize(1);
-        assertThat(emailSender.getTos().getFirst()).isEqualTo(member.getEmail());
+    @DisplayName("회원 email 중복 체크")
+    void duplicateFail() {
+        Member member = memberRegister.register(MemberFixture.createMemberRegisterReques());
+        assertThatThrownBy(()-> memberRegister.register(MemberFixture.createMemberRegisterReques())).isInstanceOf(DuplicateEmailException.class);
     }
 
-    @Test
-    @DisplayName("회원 생성 Mockito")
-    void registerTestMockito() {
-        EmailSender emailSender = Mockito.mock(EmailSender.class);
-        MemberRegister register = new MemberService(
-                new MemberRepositoryStub(), emailSender, MemberFixture.createPasswordEncoder()
-        );
-
-        Member member = register.register(MemberFixture.createMemberRegisterReques());
-        assertThat(member.getId()).isNotNull();
-        assertThat(member.getStatus()).isEqualTo(MemberStatus.PENDING);
-
-        Mockito.verify(emailSender).send(eq(member.getEmail()), any(), any());
-
-
-    }
-
-    static class MemberRepositoryStub implements MemberRepository {
-
-        @Override
-        public Member save(Member member) {
-            ReflectionTestUtils.setField(member, "id", 1L);
-            return member;
-        }
-    }
-
-
-    static class EmailSenderStub implements EmailSender {
-
-        @Override
-        public void send(Email email, String subject, String body) {
-        }
-    }
-
-    static class EmailSenderMock implements EmailSender {
-
-        List<Email> tos = new ArrayList<>();
-
-        @Override
-        public void send(Email email, String subject, String body) {
-            tos.add(email);
-        }
-
-        public List<Email> getTos() {
-            return tos;
-        }
-    }
 }
